Fix comparison between signed and unsigned integer expressions
Simple typecast to size_t should be enough to silence the warning on
ARMv7, even though the code is in fact correct, because the readlen is
checked for being < 0 in the block before the warning.

+++ lib/isc/netmgr/http.c
@@ -704,7 +704,7 @@ https_readcb(isc_nmhandle_t *handle, isc_result_t result, isc_region_t *region,
 		return;
 	}
 
-	if (readlen < region->length) {
+	if ((size_t)readlen < region->length) {
 		INSIST(session->bufsize == 0);
 		INSIST(region->length - readlen < MAX_DNS_MESSAGE_SIZE);
 		memmove(session->buf, region->base, region->length - readlen);
